Move smoothed out random to utils
